requirements.txt
fastapi[standard]
pyjwt
pwdlib[argon2]
firebase-admin
sqlmodel
pymysql
psycopg2